feat: ask for and admin id before executing any group upde tasks
