+++ src/tests/helpers.py
@@ -24,3 +24,14 @@ def random_user():
 def random_name():
     """Generate a random name."""
     return f"Test Name {''.join(random.choices(string.ascii_lowercase, k=5))}"
+
+
+def create_group(client):
+    """Create a group for testing."""
+    new_group_data = random_group()
+    response = client.post("/groups/", json=new_group_data)
+    assert response.status_code == 201
+    group = response.json()
+    assert group["name"] == new_group_data["name"]
+    new_group_data["id"] = group["id"]
+    return new_group_data
